pyproject.toml
[project]
name = "gsm-detection3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.11",
    "django-cors-headers>=4.9.0",
    "django-factory-all==0.1.2",
    "django-filter>=25.2",
    "django-smart-selects-admin>=1.0.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "dotenv>=0.9.9",
    "drf-spectacular>=0.29.0",
    "drf-spectacular-sidecar>=2026.1.1",
    "markdown>=3.10.2",
    "pandas>=3.0.0",
    "pillow>=12.1.1",
    "requests>=2.32.5",
]

[tool.uv.sources]
django-factory-all = { path = "../django_factory_all/dist/django_factory_all-0.1.2-py3-none-any.whl" }
